+++ include/seqan3/alignment/matrix/detail/trace_matrix_full.hpp
@@ -22,8 +22,8 @@
 #include <seqan3/alignment/matrix/trace_directions.hpp>
 #include <seqan3/core/detail/template_inspection.hpp>
 #include <seqan3/range/container/aligned_allocator.hpp>
-#include <seqan3/range/views/repeat_n.hpp>
 #include <seqan3/utility/detail/exposition_only_concept.hpp>
+#include <seqan3/utility/views/repeat_n.hpp>
 #include <seqan3/utility/views/zip.hpp>
 
 namespace seqan3::detail
