+++ include/seqan3/io/sam_file/detail/format_sam_base.hpp
@@ -37,7 +37,6 @@
 #include <seqan3/io/sequence_file/output_options.hpp>
 #include <seqan3/range/detail/misc.hpp>
 #include <seqan3/range/views/istreambuf.hpp>
-#include <seqan3/range/views/repeat_n.hpp>
 #include <seqan3/range/views/slice.hpp>
 #include <seqan3/range/views/take_until.hpp>
 #include <seqan3/range/views/to.hpp>
@@ -45,6 +44,7 @@
 #include <seqan3/utility/detail/exposition_only_concept.hpp>
 #include <seqan3/utility/detail/type_name_as_string.hpp>
 #include <seqan3/utility/tuple/concept.hpp>
+#include <seqan3/utility/views/repeat_n.hpp>
 #include <seqan3/utility/views/zip.hpp>
 
 namespace seqan3::detail
