+++ include/seqan3/range/container/small_vector.hpp
@@ -21,8 +21,8 @@
 
 #include <seqan3/core/concept/cereal.hpp>
 #include <seqan3/core/detail/template_inspection.hpp>
-#include <seqan3/range/views/repeat_n.hpp>
 #include <seqan3/utility/detail/integer_traits.hpp>
+#include <seqan3/utility/views/repeat_n.hpp>
 
 namespace seqan3
 {
