+++ test/snippet/utility/views/repeat_n.cpp
@@ -1,8 +1,8 @@
+#include <seqan3/std/ranges>
 #include <string>
 
 #include <seqan3/core/debug_stream.hpp>
-#include <seqan3/range/views/repeat_n.hpp>
-#include <seqan3/std/ranges>
+#include <seqan3/utility/views/repeat_n.hpp>
 
 int main()
 {
