+++ test/unit/utility/views/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_subdirectories()
 
+seqan3_test(repeat_n_test.cpp)
 seqan3_test(repeat_test.cpp)
 seqan3_test (zip_test.cpp)
